Yet another code optimization. This one really doesn't hurt.
git-svn-id: https://subversion.assembla.com/svn/sdrsharp/trunk@952 d3cde5d6-82b8-4669-8b62-43054d6781d8

+++ Radio/Decimator.cs
@@ -665,7 +665,7 @@ namespace SDRSharp.Radio
             {
                 var even = buffer[i];
                 var odd = buffer[i + 2];
-                buffer[j] = (float) (0.125 * (odd + _xEven + 3.0 * (_xOdd + even)));
+                buffer[j] = 0.125f * (odd + _xEven + 3.0f * (_xOdd + even));
                 _xOdd = odd;
                 _xEven = even;
             }
